Custom formatting issue
Issue: We are trying to use custom formatter.xml for formatting. But we
are not getting the expected formatting.

RCA: We have debugged the writer EclipseWriter.java and found that
properties defined in the formatter.xml file, are not reading in
getOptionsFromConfigFile file.

Fix: We have found that the variable “loadProfile” is used in for loop
condition. Since its default value is “false” it will not enter into the
loop. We changed the condition to “!loadProfile”.

*Inside the loop to prevent loading multiple PROFILES, “loadProfile” is
set to true.

Also while adding formatting properties to the “options” map used “id”
and “value” key for adding property “name” and “value”.  Here the id and
value will be overwritten for every iteration and latest property name
and value will be return from getOptionsFromConfigFile.

We fixed this also.