Removing unused code
We have found that few of the code are not been used.

The instance variable "overrideConfigCompilerVersion" has default value
"false". It is used in line no 117 as
"Boolean.TRUE.equals(overrideConfigCompilerVersion)", it never be true.

Also the values are added to options map in line 110-112 is not
required.

If if (configFile != null) in line 114 is true,
The values are in options map will be overwritten at line 122, since all
the values are come from formatter.xml.

If if (configFile != null) in line 114 is false,
Then at line 71 return null.

Hence the default values added in options map is removed and added check
!options.isEmpty() before line 71.